Make GetUnitFrame return target/player/focus/targettarget even with default options
Since `defaultOptions` has `ignoreTargetFrame = true` (and so on), calling `GetUnitFrame("target")`, without specifying `ignoreTargetFrame = false` in additional options, returns nothing.
Same for player/focus/targettarget.

This is a bit unfortunate and counterintuitive I think, especially for WeakAuras users.
When trying to anchor to these frames with the built in "anchored to: unit frames", it simply doesn't work because it calls `GetUnitFrame` without the additional options. This *might* be better fixed on WAs end? But I still think that even outside of WA, I'd like to get my target frame when simply doing `GetUnitFrame("target")`, without the need for `GetUnitFrame("target", {ignoreTargetFrame = true})